+++ src/bruteforce_cuda.cu
@@ -4,8 +4,8 @@
 
 __global__ void lib_kernel() { printf("WHY IS THIS NOT WORKING"); }
 
-uint64_t BruteforceCUDA::solve(Graph *graph,
-                               std::optional<uint64_t> entry_value) {
+uint16_t BruteforceCUDA::solve(Graph *graph,
+                               std::optional<uint16_t> entry_value) {
   lib_kernel<<<1, 1>>>();
   return 0;
 }
